Move code to `diagnostics.rs`
